No functional changes other than a basic start at using pthreads.  In reality this doesn't really do threading properly yet, but the packets are processed by a separate thread.  However....
There is no proper queue or scheduling yet, so in all likelihood this will be
slower than the master branch.  So really... unless you feel like doing
development work with pthreads, avoid this for now ;)
