+++ fingerprintls/Makefile
@@ -1,6 +1,6 @@
 # Commented this out for now, not sure how many use klang
 #CC=g++
-CFLAGS=-Wall -pedantic
+CFLAGS=-Wall -pedantic -std=c11 -Os 
 LDFLAGS=-lpcap
 
 all: fingerprintls
